+++ crates/red_knot_python_semantic/src/types.rs
@@ -4903,6 +4903,60 @@ impl<'db> Type<'db> {
             | Self::AlwaysFalsy => None,
         }
     }
+
+    /// Returns a tuple of two spans. The first is
+    /// the span for the identifier of the function
+    /// definition for `self`. The second is
+    /// the span for the parameter in the function
+    /// definition for `self`.
+    ///
+    /// If there are no meaningful spans, then this
+    /// returns `None`. For example, when this type
+    /// isn't callable.
+    ///
+    /// When `parameter_index` is `None`, then the
+    /// second span returned covers the entire parameter
+    /// list.
+    ///
+    /// # Performance
+    ///
+    /// Note that this may introduce cross-module
+    /// dependencies. This can have an impact on
+    /// the effectiveness of incremental caching
+    /// and should therefore be used judiciously.
+    ///
+    /// An example of a good use case is to improve
+    /// a diagnostic.
+    fn parameter_span(
+        &self,
+        db: &'db dyn Db,
+        parameter_index: Option<usize>,
+    ) -> Option<(Span, Span)> {
+        match *self {
+            Type::FunctionLiteral(function) => {
+                let function_scope = function.body_scope(db);
+                let span = Span::from(function_scope.file(db));
+                let node = function_scope.node(db);
+                let func_def = node.as_function()?;
+                let range = parameter_index
+                    .and_then(|parameter_index| {
+                        func_def
+                            .parameters
+                            .iter()
+                            .nth(parameter_index)
+                            .map(|param| param.range())
+                    })
+                    .unwrap_or(func_def.parameters.range);
+                let name_span = span.clone().with_range(func_def.name.range);
+                let parameter_span = span.with_range(range);
+                Some((name_span, parameter_span))
+            }
+            Type::BoundMethod(bound_method) => {
+                Type::FunctionLiteral(bound_method.function(db)).parameter_span(db, parameter_index)
+            }
+            _ => None,
+        }
+    }
 }
 
 impl<'db> From<&Type<'db>> for Type<'db> {
